extracted BlProxy from BlRequest.h and prepared a renaming to ComProxy
