+++ BlRequest.h
@@ -19,7 +19,6 @@
 #ifndef _BL_REQUEST_H_
 #define _BL_REQUEST_H_
 
-#include "ClientSocket.h"
 #include <cassert>
 #include <cstring>
 #include <stdio.h>
@@ -86,27 +85,6 @@ struct BlReadRequest : public BlAddressRequest
 
 		unsigned char numBytesLow;
 		unsigned char numBytesHigh;
-};
-
-class BlProxy
-{
-	private:
-		const ClientSocket* const mSock;
-
-	public:
-		BlProxy(const ClientSocket* const pSock);
-
-		/**
-		 * Mask bytes of input buffer and add CRC16-CITT checksum to the end
-		 * @param pInput input buffer
-		 * @param inputLength number of bytes in input buffer
-		 * @param pOutput output buffer
-		 * @param outputLength size of the output buffer
-		 */
-		size_t MaskControlCharacters(const unsigned char* pInput, size_t inputLength, unsigned char* pOutput, size_t outputLength) const;
-		int Send(BlRequest& req, unsigned char* pResponse, size_t responseSize, bool doSync = true) const;
-		int Send(const unsigned char* pRequest, const size_t requestSize, unsigned char* pResponse, size_t responseSize, bool checkCrc, bool sync) const;
-		size_t UnmaskControlCharacters(const unsigned char* pInput, size_t inputLength, unsigned char* pOutput, size_t outputLength, bool checkCrc) const;
 };
 
 struct BlInfo
