+++ ComProxy.cpp
@@ -16,7 +16,7 @@
  You should have received a copy of the GNU General Public License
  along with Wifly_Light.  If not, see <http://www.gnu.org/licenses/>. */
 
-#include "BlRequest.h"
+#include "ComProxy.h"
 #include "crc.h"
 #include "trace.h"
